feat: MAT-269 선수 할당 기능 어댑터 분리 - 교체 리스트
