+++ src/routes/matches/$matchId/players/edit/-components/PlayersEditView.tsx
@@ -62,6 +62,7 @@ export const PlayersEditView = ({ matchId }: { matchId: number }) => {
         <div className={styles.subListContainer}>
           <SubPlayerListForEdit
             matchId={matchId}
+            team={homeTeamWithoutTeamColor}
             players={matchPlayers.data.homeTeam.substitutes}
           />
         </div>
